Update superstyles.css
In order for Safari and Firefox to behave similar to Chrome, where the non-visible <sc-view>'s move out of the way to allow the visible <sc-view> to take up the whole area, `display:block;` and `overflow:hidden;` must be added.  For IE to behave the same `zoom: 1;` must be added.